fix: moved the custom pyvisa classes to their own files
